move contexttype to main
